Arenas no-longer generate in a rectangular shape
Arenas now always generate in the shape of a square, even when they
consist of rectangular bricks.
(A side effect of this is that 4x8 brick arenas are huge now. This will
be addressed soon.)

Rectangular arenas will be re-added later as one of the arena shape
types.

More work has been done on shaped arenas. However, it is disabled for
now because it is not working properly yet.